+++ src/app/features/log-in-page/log-in/log-in.component.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/variables.scss' as *;
-@use '../../../styles/mixings.scss' as *;
+@use '../../../../styles/variables.scss' as *;
+@use '../../../../styles/mixings.scss' as *;
 
 :host {
     height: 100%;
